refactoring math mode
